scripts

+++ Vagrantfile
@@ -21,7 +21,7 @@ VAGRANT_RAM = ENV['VAGRANT_RAM'] || "2048"
 VAGRANT_CORES = ENV['VAGRANT_CORES'] || "4"
 VAGRANT_BOXNAME = ENV['VAGRANT_BOXNAME'] || "docker1"
 VAGRANT_ANNOTATION = ENV['VAGRANT_ANNOTATION'] || "docker host"
-
+§
 # You may also provide a comma-separated list of ports
 # for Vagrant to forward. For example:
 # $ FORWARD_PORTS=8080,27017 vagrant [up|reload]
@@ -36,6 +36,25 @@ user="$1"
 if [ -z "$user" ]; then
     user=vagrant
 fi
+
+if [ ! -d "/vagrant" ]; then
+  echo "MISSING /vagrant"; 
+  exit 1
+fi
+if [ ! -d "/vagrant/data" ]; then
+  echo "MISSING: /vagrant/data"; 
+  exit 1
+fi
+if [ ! -f "/data/.empty" || ! -f "/vagrant/data/.empty" ]; then
+  echo "MISSING .empty file in /data and /vagrant"
+  exit 1
+fi
+
+mount
+ls -la /
+ls -la /vagrant
+ls -la /data
+
 
 
 [[ -f /etc/apt/apt.conf.d/01proxy ]] && rm -v /etc/apt/apt.conf.d/01proxy
@@ -104,7 +123,14 @@ usermod -a -G docker "$user"
     ip addr
     ifconfig
 
+    /vagrant/scripts/write-interface-addresses.sh \
+        /data/state \
+        eth0 \
+        docker 0
+
+    echo "pre-reboot"
     shutdown -r now
+
 
     echo "after-reboot"
 ##fi
@@ -138,6 +164,7 @@ Vagrant::Config.run do |config|
   # Setup virtual machine box. This VM configuration code is always executed.
   config.vm.box = BOX_NAME
   config.vm.box_url = BOX_URI
+  config.vm.hostname = VAGRANT_BOXNAME
 
   # Use the specified private key path if it is specified and not empty.
   if SSH_PRIVKEY_PATH
